Use size_t for token buffer length in picoquic_prepare_retry_token

+++ picoquic/packet.c
@@ -1056,11 +1056,11 @@ int picoquic_incoming_initial(
         if (picoquic_verify_retry_token((*pcnx)->quic, addr_from, current_time,
             &(*pcnx)->original_cnxid, ph->token_bytes, ph->token_length) != 0) {
             uint8_t token_buffer[256];
-            uint32_t token_size;
+            size_t token_size;
 
             if (picoquic_prepare_retry_token((*pcnx)->quic, addr_from, 
                 current_time + PICOQUIC_TOKEN_DELAY_SHORT, &ph->dest_cnx_id,
-                token_buffer, (uint32_t)sizeof(token_buffer), &token_size) != 0){ 
+                token_buffer, sizeof(token_buffer), &token_size) != 0){ 
                 ret = PICOQUIC_ERROR_MEMORY;
             }
             else {
